# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ultra Computer API
  version: 0.9.0
  description: |
    Full REST API for Ultra Computer — an AI orchestration platform with multi-agent
    swarm coordination, autonomous scheduling, voice, browser automation, marketplace,
    identity, messaging, and protocol interoperability (A2A, MCP, CLI, webhooks).
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
paths:
...
  /api/setup/configure:
    post:
      operationId: configureSetup
      summary: Apply setup configuration (provider keys, model assignments, etc.)
      tags: [Setup]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                apiKey:
                  type: string
                modelId:
                  type: string
                role:
                  type: string
      responses:
        '200':
          description: Configuration applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  message:
                    type: string
        '400':
          description: Configuration error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/setup/test-connection:
    post:
      operationId: testSetupConnection
      summary: Test a provider connection during setup
      tags: [Setup]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider]
              properties:
                provider:
                  type: string
                apiKey:
                  type: string
                baseUrl:
                  type: string
      responses:
        '200':
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  latencyMs:
                    type: integer
                  models:
                    type: array
                    items:
                      type: string
                  error:
                    type: string
        '400':
          description: Test failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── API Docs (Swagger UI) ───────────────────────────────────────────────────
...